feat(terminal): rejoin width-wrapped lines when saving to file
Extract the terminal-buffer extraction into bufferToLogicalLines and add a
gated full-width-join mode: rows that completely fill the terminal width are
rejoined with the following row, so lines a program hard-wrapped at the
terminal width are saved as single logical lines. Deliberate blank lines are
preserved. Save enables this mode; copy-to-clipboard keeps the existing
soft-wrap-only behavior.
